LayoutTests/platform/qt/fast/js/constructor-length-expected.txt

This tests the length property of constructors.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS ArrayBuffer.length is 1
FAIL AudioContext.length should be 0. Threw exception ReferenceError: Can't find variable: AudioContext
PASS BeforeLoadEvent.length is 2
PASS Blob.length is 2
PASS CloseEvent.length is 2
PASS CustomEvent.length is 2
FAIL DOMFormData.length should be 0. Threw exception ReferenceError: Can't find variable: DOMFormData
PASS DOMParser.length is 0
PASS DataView.length is 3
PASS ErrorEvent.length is 2
PASS Event.length is 2
PASS EventSource.length is 1
PASS Float32Array.length is 1
PASS Float64Array.length is 1
PASS FileReader.length is 0
FAIL FileReaderSync.length should be 0. Threw exception ReferenceError: Can't find variable: FileReaderSync
PASS HashChangeEvent.length is 2
PASS Int16Array.length is 1
PASS Int32Array.length is 1
PASS Int8Array.length is 1
FAIL Intent.length should be 3. Threw exception ReferenceError: Can't find variable: Intent
PASS MediaController.length is 0
FAIL MediaStream.length should be 2. Threw exception ReferenceError: Can't find variable: MediaStream
PASS MessageChannel.length is 0
PASS MessageEvent.length is 2
PASS OverflowEvent.length is 2
PASS PageTransitionEvent.length is 2
FAIL PeerConnection.length should be 2. Threw exception ReferenceError: Can't find variable: PeerConnection
PASS PopStateEvent.length is 2
PASS ProgressEvent.length is 2
FAIL ShadowRoot.length should be 1. Threw exception ReferenceError: Can't find variable: ShadowRoot
PASS SharedWorker.length is 2
PASS StorageEvent.length is 2
FAIL TextTrackCue.length should be 6. Threw exception ReferenceError: Can't find variable: TextTrackCue
FAIL TrackEvent.length should be 2. Threw exception ReferenceError: Can't find variable: TrackEvent
PASS Uint16Array.length is 1
PASS Uint32Array.length is 1
PASS Uint8Array.length is 1
PASS Uint8ClampedArray.length is 1
PASS WebGLContextEvent.length is 2
PASS WebKitAnimationEvent.length is 2
PASS WebKitCSSMatrix.length is 1
FAIL WebKitMutationObserver.length should be 1. Threw exception ReferenceError: Can't find variable: WebKitMutationObserver
PASS WebKitPoint.length is 2
PASS WebKitTransitionEvent.length is 2
PASS WebSocket.length is 1
PASS Worker.length is 1
PASS URL.length is 0
PASS XMLHttpRequest.length is 0
PASS XMLSerializer.length is 0
PASS XPathEvaluator.length is 0
PASS XSLTProcessor.length is 0
PASS successfullyParsed is true

TEST COMPLETE
